Extract VPalm parameters

Format the model outputs as a list for VPalm inputs

extract_params(data, model, leaves = 45, seed = sample.int(1000, 1),
  init = init_list(), type = c("sample", "mean"))

Arguments

data
A list of all data (generally from import_data())

model
A list of the models fitted on the data (generally from mod_all())

leaves
The number of leaves needed

seed
The seed for random parameter generation

init
The initialisation parameters (see details)

type
The type of sampling performed on the parameter distribution (see coef_sample())

Value

A list of the parameters used as VPalm input

Details

Parameters sampled include:

The rachis curvature

The rachis relative height

All lme models (see models)

All nlme models (see models) The type parameter controls whether a sample or an average tree is needed.

The init parameter should be a named list of initialization parameters as in init_list(). The user can use init_list() as a template to change the values of the init argument (see examples).

Examples

# Change the initial values of lamina angle :
ini= init_list()
ini$laminaAngle= 145
# NOT RUN {
 # call the function to create a VPalm input:
 extract_params(data,model, leaves= 45, seed= sample.int(1000,1),
 initialisation= ini)
# }