Compute the leaflet angles

Estimate the leaflet angles at point A and C using the tangents along the leaf curvature

find_angle(x, y, res = 1, method = "smooth.spline", ...)

Arguments

x
The values in the x axis

y
The values in the y axis

res
The re-sampling resolution, with the same unit as inputs.

method
The smoothing method, either using stats::loess() or stats::smooth.spline()

...
Further arguments to pass to the smoothing function

Value

The tangent for each point of a curved object

Details

The x and y coordinates must be ordered from first to last point on the object. The function is used to compute the tangents of points for a curved object. The function re-sample the object using a smoothing function to give the closest tangeant to the point. The higher the resolution, the higher the sub-sampling, and the closer the tangent will be. Here is how the function works:

The user sample some points from a curved object and register their X and Y positions in a euclidean coordinate system, and give the points coordinates as input of the function

An approximation of the object curvature is made using a smoothing function on all measured points

The object is re-sampled using the resulting smoothing function with a sampling resolution of res.

The tangent is computed for each input point using the closer re-sampled points

See also

Examples

if (FALSE) {
library(Vpalmr)
find_angle(x= 1:10,y= rep(1,10))
}