Import all data

Import and pre-compute the input data for VPalm

import_data(parameter, development, phylotaxy, declination, curvature, toricity,
  leaf_area, axial_angle, petiole_width, twist)

Arguments

parameter
File path to the parameter file (ParameterSimu.csv)

development
File path to the development file (Development_Rep4_SMSE.csv)

phylotaxy
File path to the phylotaxy file (Stem_SMSE14.csv)

declination
File path to the declination file (A_SMSE_Nov14.csv)

curvature
File path to the leaf curvature file (LeafCurvature_SMSE14.csv)

toricity
File path to the leaf toricity file (AnglesC&A_SMSE_Nov14.csv)

leaf_area
File path to the leaf area file (LeafArea_monitoring_SMSE.csv)

axial_angle
File path to the leaf axial angle file (LeafDispositionComp_SMSE14.csv)

petiole_width
File path to the leaf petiole width file (Petiole_SMSE14.csv)

twist
File path to the leaf twist file (Torsion_SMSE14.csv), used to compute leaf nerve height.

Value

A list of the imported and pre-processed files

Examples

# NOT RUN {
Vpalmr::import_data(parameter= 'Archi/ParameterSimu.csv',
                    development= 'Archi/Development_Rep4_SMSE.csv',
                    phylotaxy= "Archi/Stem_SMSE14.csv",
                    declination= "Archi/AnglesC&A_SMSE_Nov14.csv",
                    curvature= "Archi/LeafCurvature_SMSE14.csv",
                    toricity= 'Archi/AnglesC&A_SMSE_Nov14.csv',
                    leaf_area= 'Archi/LeafArea_monitoring_SMSE.csv',
                    axial_angle= "Archi/LeafDispositionComp_SMSE14.csv",
                    petiole_width= "Archi/Petiole_SMSE14.csv",
                    twist= "Archi/Torsion_SMSE14.csv")
# }