Fit the models

These functions fit a given model from which the user can extract parameters used as VPalm inputs

mod_stem_diameter(data)

mod_stem_height(data)

mod_rachis_length(data, control = nlme::lmeControl(maxIter = 5e+05,
  niterEM = 25000))

mod_petiole_ratio(data)

mod_B_position(data)

mod_nb_leaflet(data)

mod_C_declination(data, control = nlme::lmeControl(maxIter = 5e+05,
  niterEM = 25000))

mod_leaf_curvature(data)

mod_A_declination(data, decMaxA = 180, decSlopeA = 0.01)

mod_A_deviation(data)

mod_leaflet_position(data, Area, control = stats::nls.control(maxiter =
  5e+05))

mod_leaflet_length_B(data, control = nlme::lmeControl(maxIter = 5e+05,
  niterEM = 25000))

mod_leaflet_width_B(data, control = nlme::lmeControl(maxIter = 5e+05,
  niterEM = 25000))

mod_leaflet_length(data, control = nlme::nlmeControl(maxIter = 5e+05,
  niterEM = 25000))

mod_leaflet_width(data, control = nlme::nlmeControl(maxIter = 5e+05,
  niterEM = 25000))

mod_leaflet_axial_angle(data, control = nlme::nlmeControl(maxIter =
  5e+05, niterEM = 25000))

mod_leaflet_radial_angle(data)

mod_leaflet_type_freq(data)

mod_leaflet_shape(data)

mod_petiole_width_C(data, control = nlme::lmeControl(maxIter = 5e+05,
  niterEM = 25000))

mod_petiole_height(data, control = nlme::nlmeControl(maxIter = 5e+05,
  niterEM = 25000))

Arguments

data
Input data, generally one of the objects from import_data()

control
The control parameters for each kind of model

decMaxA
The maximum allowed declination angle at the A point

decSlopeA
The declination slope at point A

Area
The area data from import_data()

Value

A tibble::tibble() of the model outputs, eventually grouped by Progeny if several in input

Purpose of each function

mod_stem_diameter(): model the stem diameter from the rachis length using a sigmoid()

mod_stem_height(): model the stem height using the total emitted leaves throughout the palm life using stem_height()

mod_rachis_length(): model the leaf length from the leaf number (index of the leaves starting from the first leaf ever emitted)

mod_petiole_ratio(): model the petiole/rachis ratio by taking the average value for the progeny at the given physiological age +/- 60. If there is no value, return the average value by progeny for all leaves older than the reference leaf rank (>17).

mod_nb_leaflet(): model the number of leaflets from the rachis length using a sigmoid()

mod_C_declination(): model the leaf declination angle at C point from leaf rank

mod_leaf_curvature(): find the leaf inclination parameters (coefCurv and decAInfl) by minimizing the difference between a computed leaf using these parameters and the measurements. Uses optim_Leaf_Curv_Infl() under the hood.

mod_A_declination(): model the leaf declination angle at A point from C angle. This function needs the decAInfl parameter computed by mod_leaf_curvature(), so it first checks if the parameter is present in the input data, and computes it if absent. The user should avoid this step by providing the parameter when possible because the mod_leaf_curvature() call can be computationally intensive.

mod_A_deviation(): model the leaf deviation by using the distance from C point and the length of the leaf between A and C point as inputs for deviation_A(), and return the average value and standard deviation per progeny.

mod_leaflet_position(): model the leaflet position along the leaf rachis.

mod_leaflet_length_B(): model the leaflet length at B point

mod_leaflet_width_B(): model the leaflet width at B point

mod_leaflet_length(): model the leaflet relative length from its relative position on the rachis

mod_leaflet_width(): model the leaflet relative width from its relative position on the rachis

mod_leaflet_axial_angle(): model the leaflet axial angle from its relative position on the rachis

mod_leaflet_radial_angle(): model the leaflet radial angle from its relative position on the rachis

mod_leaflet_type_freq(): model the leaflet radial angle frequency along the rachis. Leaflets can be positionned up, medium or down.

mod_leaflet_shape(): model the leaflet shape

mod_petiole_width_C(): model the petiole width at C point

mod_petiole_height(): model the petiole section height using section_height()

The models

stats::nls() is used for mod_stem_diameter(), mod_stem_height(), mod_nb_leaflet(), mod_leaflet_position()

nlme::lme() is used for mod_rachis_length(), mod_C_declination(), mod_leaflet_length_B(), mod_petiole_width_C(), mod_petiole_height()

nlme::nlme() is used for mod_leaflet_length(), mod_leaflet_width(), mod_leaflet_axial_angle()

base::mean() is used for mod_petiole_ratio()

stats::optim() is used by mod_leaf_curvature()