Plot the architectural parameters estimations

Plot the estimations of the palm architectural parameters from their parameter estimation and input data.

plot_stem_diameter(data, model)

plot_stem_height(data, model)

plot_rachis_length(data, model, nb_leaves)

plot_petiole_ratio(data, model)

plot_B_position(data, model)

plot_nb_leaflet(data, model)

plot_C_declination(data, model)

plot_leaf_curvature(data, model)

plot_A_declination(data, model, decMaxA = 180, decSlopeA = 0.01)

plot_A_deviation(data, model)

plot_leaflet_position(data, model, mode = c("Progeny", "Tree"))

plot_leaflet_length_B(data, model)

plot_leaflet_width_B(data, model)

plot_leaflet_length(data, model)

plot_leaflet_width(data, model)

plot_leaflet_axial_angle(data, model)

plot_leaflet_radial_angle(data, model)

plot_leaflet_type_freq(model)

plot_leaflet_shape(model)

plot_petiole_width_C(data, model)

plot_petiole_height(data, model)

Arguments

data
Input data, generally one of the objects from import_data()

model
Model fitted from one of models

nb_leaves
Number of leaves required for model prediction

decMaxA
The maximum allowed declination angle at the A point

decSlopeA
The declination slope at point A

mode
Information level of the plot: "Progeny" for progeny scale, or "Tree" for tree scale.

Value

A ggplot

Plotting functions (plot_*)

plot_stem_diameter(): plot the predicted stem diameter at the level of the considered leaf using the leaf length

plot_stem_height(): plot the predicted stem height using the total cumulated emitted leaves

plot_rachis_length(): plot the predicted leaf length according to its index (leaf number)

plot_petiole_ratio(): plot the observed petiole to rachis ratio along the average value used for simulation.

plot_nb_leaflet(): plot the predicted number of leaflets in function of rachis length

plot_C_declination(): plot the declination angle at C point in function of leaf rank

plot_leaf_curvature(): plot the leaf curvature in function of the relative position on the leaf

plot_A_declination(): plot the leaf declination angle at A point in function of the C point

plot_A_deviation(): plot the leaf deviation computed using deviation_A(), compared to the value used by Vpalm (the average value).

plot_leaflet_position(): plot the leaflet position along the rachis relative rank for each progeny with mode= "Progeny" or for each Progeny and colored by Tree with mode= "Tree"

plot_leaflet_length_B(): plot the predicted leaflet length at B point

plot_leaflet_width_B(): plot the predicted leaflet width at B point

plot_leaflet_length(): plot the predicted relative leaflet length

plot_leaflet_width(): plot the predicted leaflet width

plot_leaflet_axial_angle(): plot the predicted leaflet axial angle from the relative position of the leaflet on the rachis

plot_leaflet_radial_angle(): plot the leaflet radial angle from its relative position on the rachis

plot_leaflet_type_freq(): plot the leaflet radial angle frequency along the rachis. Leaflets can be positionned up, medium or down.

plot_leaflet_shape(): plot the leaflet shape

plot_petiole_width_C(): plot the petiole width at C point

plot_petiole_height(): plot the petiole heigth

Examples

# NOT RUN {
library(Vpalmr)
archi_param= estimate_archi(progeny= "DY",map= 47)
plot_params(archi_param)
# }