VPalm parameter list for all progenies

Make a VPalm parameter list for n trees sampled in the parameters distributions of each progeny, and also for an average tree from the n trees distribution. provided in the data from the models outputs using extract_params() and replace missing value by the average of all progenies.

extract_progeny(data, model, n, leaves = 45, seed = NULL,
  init = init_list(), average = T)

Arguments

data
A list of all data (generally from import_data())

model
A list of the models fitted on the data (generally from mod_all())

n
The number of samplings for each progeny, or in other words, the number of trees to be generated for each progeny

leaves
The number of leaves needed

seed
The seeds for random parameter generation, see details.

init
Initialisation values (see extract_trees())

average
Boolean, is mean of all average trees to be used to replace missing values ?

Value

A nested list, containing a list of VPalm parameters for each tree sampled for each Progeny, and the average tree of each progeny, and one with the average (non-NA) values for all progenies. Each list in the list has as many n lists that represents each sampled tree for each progeny.

Details

seed should be a named list of length equal to the number of progenies, and each object with n seeds each. The names of the list objects should match the ones of the progenies. Alternatively, seed can be a vector of seeds of length n that will be recycled for each progeny. If NULL, a random seed is generated for each tree of each progeny using base::sample().