Extract sheets

Extract the data from all forms and all sheets from the leaf area/biomass or the development excel files

extract_sheets(path, form = c("area", "development"), sheet = NULL)

Arguments

path
The path to the excel file

form
The form type (see form section)

sheet
The sheets index on the files (see details)

Value

A data.frame with the data from all sheets (all leaves)

Details

If the sheet is not given, the function will import data from all sheets. Otherwise, the user can provide a given sheet, or a vector of sheets (see examples).

Note

If you have an error as the following one: "JAVA_HOME cannot be determined from the Registry", please check that you have a JAVA version that match you R one, i.e. that your JAVA is 64-bits if your R is 64-bits also. If JAVA is 64-bits, it should be installed in "C:/Program Files/Java", otherwise it is located in "C:/Program Files (x86)/Java"

Examples

# NOT RUN {
# Extract all sheets:
extract_sheets(path= "1-Data/raw/Form Archi.xlsx",form='area')

# Extract only the first sheet:
form(path= "1-Data/raw/Form Archi.xlsx", form='area', sheet= 1)

# Extract the second and third sheets:
extract_sheets(path= "1-Data/raw/Form Archi.xlsx",form='area', sheet= c(2,3))
# }